+++ src/components/TextInput/TextInput.tsx
@@ -2,14 +2,17 @@ import {
   chakra,
   Input as ChakraInput,
   Textarea as ChakraTextarea,
+  useMergeRefs,
   useMultiStyleConfig,
 } from "@chakra-ui/react";
-import React, { forwardRef } from "react";
+import React, { forwardRef, useEffect, useRef, useState } from "react";
 
 import ComponentWrapper from "../ComponentWrapper/ComponentWrapper";
 import Label from "../Label/Label";
 import { HelperErrorTextType } from "../HelperErrorText/HelperErrorText";
 import { getAriaAttrs } from "../../utils/utils";
+import Button from "../Button/Button";
+import Icon from "../Icons/Icon";
 
 // HTML Input types as defined by MDN: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input
 export type TextInputTypes =
@@ -47,6 +50,8 @@ export interface InputProps {
   id: string;
   /** Populates the HelperErrorText for the error state */
   invalidText?: HelperErrorTextType;
+  /** Adds a button to clear existing text in the input field. */
+  isClearable?: boolean;
   /** Adds the `disabled` and `aria-disabled` prop to the input when true */
   isDisabled?: boolean;
   /** Adds errored styling to the input/textarea and helper text elements */
@@ -77,6 +82,9 @@ export interface InputProps {
   onFocus?: (event: React.FocusEvent<HTMLInputElement>) => void;
   /** Populates the placeholder for the input/textarea elements */
   placeholder?: string;
+  /** Allows the '(Required)' text to be changed for language purposes
+   * Note: Parenthesis will be added automatically by the component */
+  requiredLabelText?: string;
   /** Offers the ability to hide the helper/invalid text. */
   showHelperInvalidText?: boolean;
   /** Offers the ability to show the label onscreen or hide it. Refer to the
@@ -85,9 +93,6 @@ export interface InputProps {
   /** Whether or not to display the "(Required)" text in the label text.
    * True by default. */
   showRequiredLabel?: boolean;
-  /** Allows the '(Required)' text to be changed for language purposes
-   * Note: Parenthesis will be added automatically by the component */
-  requiredLabelText?: string;
   /** The amount to increase or decrease when using the number type. */
   step?: number;
   /** FOR INTERNAL DS USE ONLY: the input variant to display. */
@@ -118,6 +123,7 @@ export const TextInput = chakra(
         helperText,
         id,
         invalidText,
+        isClearable = false,
         isDisabled = false,
         isInvalid = false,
         isRequired = false,
@@ -140,7 +146,13 @@ export const TextInput = chakra(
         value,
         ...rest
       } = props;
+      const [finalValue, setFinalValue] = useState<string>(value || "");
+      const closedRef = useRef<HTMLInputElement>();
+      const mergedRefs = useMergeRefs(closedRef, ref);
+      // If a ref is not passed, then merging refs won't work.
+      const finalRef = ref ? mergedRefs : closedRef;
       const styles = useMultiStyleConfig("TextInput", {
+        showLabel,
         variant: textInputType,
       });
       const isTextArea = type === "textarea";
@@ -148,6 +160,14 @@ export const TextInput = chakra(
       const finalInvalidText = invalidText
         ? invalidText
         : "There is an error related to this field.";
+      const internalOnChange = (
+        e:
+          | React.ChangeEvent<HTMLInputElement>
+          | React.ChangeEvent<HTMLTextAreaElement>
+      ) => {
+        setFinalValue(e.target.value);
+        onChange && onChange(e);
+      };
       let footnote: HelperErrorTextType = isInvalid
         ? finalInvalidText
         : helperText;
@@ -158,9 +178,21 @@ export const TextInput = chakra(
         name: "TextInput",
         showLabel,
       });
+      const onClearClick = () => {
+        setFinalValue("");
+        // Set focus back to the input element.
+        (finalRef as any).current.focus();
+      };
       let finalIsInvalid = isInvalid;
       let fieldOutput;
+      let clearButtonOutput;
       let options;
+
+      useEffect(() => {
+        if (value && value !== finalValue) {
+          setFinalValue(value);
+        }
+      }, [finalValue, value]);
 
       if (!id) {
         console.warn(
@@ -188,7 +220,14 @@ export const TextInput = chakra(
 
       // When the type is "hidden", the input element needs fewer attributes.
       options = isHidden
-        ? { defaultValue, id, "aria-hidden": isHidden, name, onChange, ref }
+        ? {
+            defaultValue,
+            id,
+            "aria-hidden": isHidden,
+            name,
+            onChange: internalOnChange,
+            ref: finalRef,
+          }
         : {
             "aria-required": isRequired,
             defaultValue,
@@ -200,11 +239,11 @@ export const TextInput = chakra(
             maxLength,
             min,
             name,
-            onChange,
+            onChange: internalOnChange,
             onClick,
             onFocus,
             placeholder,
-            ref,
+            ref: finalRef,
             // The `step` attribute is useful for the number type.
             step: type === "number" ? step : null,
             ...ariaAttributes,
@@ -213,9 +252,27 @@ export const TextInput = chakra(
       // For `input` and `textarea`, all attributes are the same but `input`
       // also needs `type` and `value` to render correctly.
       if (!isTextArea) {
-        options = { type, value, ...options } as any;
+        options = { type, value: finalValue, ...options } as any;
         fieldOutput = <ChakraInput {...options} __css={styles.input} />;
+        if (isClearable && !isHidden) {
+          clearButtonOutput = (
+            <Button
+              buttonType="text"
+              id={`${id}-clear-btn`}
+              onClick={onClearClick}
+              sx={styles.clearButton}
+            >
+              <Icon color="ui.black" name="close" size="medium" />
+              <span>Clear {labelText}</span>
+            </Button>
+          );
+        }
       } else {
+        if (isClearable) {
+          console.warn(
+            "NYPL Reservoir TextInput: The `isClearable` prop cannot be used with the `textarea` type."
+          );
+        }
         fieldOutput = (
           <ChakraTextarea {...options} __css={styles.textarea}>
             {value}
@@ -245,6 +302,7 @@ export const TextInput = chakra(
             </Label>
           )}
           {fieldOutput}
+          {!isHidden && finalValue.length > 0 && clearButtonOutput}
         </ComponentWrapper>
       );
     }